% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_Bin.r
\name{SS_Bin}
\alias{SS_Bin}
\title{SS_Bin}
\usage{
SS_Bin(alpha, beta, m, theta.decline, PI)
}
\arguments{
\item{alpha}{probability of a Type I error}

\item{beta}{probability of a Type II error}

\item{m}{number of years}

\item{theta.decline}{decline as a proportion}

\item{PI}{probability of presence}
}
\description{
Sample Size assuming Bernoulli distribution
}
