% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_NB.r
\name{SS_NB}
\alias{SS_NB}
\title{SS_NB}
\usage{
SS_NB(alpha, beta, m, Gamma, mu0, phi)
}
\arguments{
\item{alpha}{probability of a Type I error}

\item{beta}{probability of a Type II error}

\item{m}{number of years}

\item{Gamma}{proportion remaining after each year (So 0.001
indicates a 99.9\% decline). This can be a vector of proportions.}

\item{mu0}{estimated mean from sample}

\item{phi}{overdispersion parameter}
}
\description{
Sample Size assuming Negative Binomial distribution
}
