% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_Pois.r
\name{SS_Pois}
\alias{SS_Pois}
\title{SS_Pois}
\usage{
SS_Pois(alpha, beta, m, Gamma, mu0)
}
\arguments{
\item{alpha}{probability of a Type I error}

\item{beta}{probability of a Type II error}

\item{m}{number of years}

\item{Gamma}{proportion remaining after each year (So 0.001
indicates a 99.9\% decline). This can be a vector of proportions.}

\item{mu0}{estimated mean from sample}
}
\description{
SS_Pois
}
\details{
Sample Size assuming Poisson distribution
}
