% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_ZIB.r
\name{SS_ZIB}
\alias{SS_ZIB}
\title{SS_ZIB}
\usage{
SS_ZIB(alpha, beta, m, theta.decline, PI, Q, V)
}
\arguments{
\item{alpha}{probability of a Type I error}

\item{beta}{probability of a Type II error}

\item{m}{number of years}

\item{theta.decline}{decline as a proportion}

\item{PI}{probability of presence}

\item{Q}{Q}

\item{V}{V}
}
\description{
Sample Size assuming zero-inflated Binomial distribution
}
