% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_ZINb.r
\name{SS_ZINb}
\alias{SS_ZINb}
\title{SS_ZINb}
\usage{
SS_ZINb(alpha, beta, m, Gamma, mu0, mix.p, phi)
}
\arguments{
\item{alpha}{probability of a Type I error}

\item{beta}{probability of a Type II error}

\item{m}{number of years}

\item{Gamma}{proportion remaining after each year (So 0.001
indicates a 99.9\% decline). This can be a vector of proportions.}

\item{mu0}{expected Poisson count}

\item{mix.p}{mixture proportion}

\item{phi}{overdispersion parameter}
}
\description{
Sample Size assuming zero-inflated negative Binomial distribution
}
