% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_ZIP.r
\name{SS_ZIP}
\alias{SS_ZIP}
\title{SS_ZIP}
\usage{
SS_ZIP(alpha, beta, m, Gamma, mu0, mix.p)
}
\arguments{
\item{alpha}{probability of a Type I error}

\item{beta}{probability of a Type II error}

\item{m}{number of years}

\item{Gamma}{proportion remaining after each year (So 0.001
indicates a 99.9\% decline). This can be a vector of proportions.}

\item{mu0}{expected Poisson count}

\item{mix.p}{mixture proportion}
}
\description{
Sample Size assuming zero-inflated Poisson distribution
}
